support: split router into multiple files
